2 Fixes. I suck at testing stuff.
Fixed tutorials frame hiding even when you did not tell it to do so.
Fixed TalkingLessHead Module not loading. We check for the proper blizzard add-on and if it is not loaded we will go ahead and force it to be sure we get our data built and the module works.