Fix the missing actionbars in "Cycle of Life" world quest
In 8.1.5 we got a new world quest where you get a vehicleui but not a vehicle actionbar. Instead blizzard are using an overridebar. To work around this issue, we moved the vehicleui conditional to the back of the line, so overridebar, possessbar and the other bar variations will fire first, thus giving us... buttons!