Massive Fonts / Texture system change
We only support KkthnxUI font now. Textures are now set by 1 global option as well which has been moved to the general section of the config. If you want different fonts, you need to make a font folder inside the wow folder and change them this way. This is a lot better for us in the long run with changing stuff. NOTE: This is an on-going process still as there could be a lot of bugs or missing font sets or textures. Thanks.