Attempted to fix "table index is nil" error in Functions.lua
Changed the ColorClass function in Functions.lua to return a default white color if the provided class does not exist in the K.ClassColors table. This should prevent the "table index is nil" error that was occurring when the function was called with an invalid class name.

Also added a fail-safe check in the K.UnitColor function to ensure that the reaction variable is not nil before attempting to use it to get a color value from the K.Colors.reaction table or FACTION_BAR_COLORS table.

This commit should fix the error message shown in the log file.