Refine Whisper Invite feature
We have decided to load modules in a different way that involves a new code format that I find preferable. I have already begun implementing this change, but it will only apply when the core file for a module is loading more than one module related to that particular module folder.